Adds the conventional ROOT_DIR for d.rymcg.tech and fixes default URL.
The upstream docs for d.rymcg.tech inform the user to clone to:
$HOME/git/vendor/enigmacurry/d.rymcg.tech
This default path is considered to be good enough for all users.
I understand some people will still desire to clone this elsewhere,
but a standard default path seems like a good idea for new users.

Fixes the default URL shown by `make config`. The default is
`wp${INSTANCE_URL_SUFFIX}.${ROOT_DOMAIN}` which expands to
wp.example.com for the default instance, and wp-foo.example.com for
the foo instance. Obviously not eveyone wants `wp` in their url, but
there should be some kind of default here. This is how its done in the
rest of the d.ry apps at least.