Adds healthcheck
You are supposed to be able to get the Traefik file provider to reload from /config/dynamic/acme-sh-certs.yml simply by touching it.
That mostly works, except that weirdly it does not seem to reload the tls settings without a full traefik restart.
So, as a workaround, I have rigged the Traefik entrypoint to be restart automatically if it finds the sentinel file at /data/restart_me (aka. /traefik/restart_me in the acme-sh container)